Feni Mirin
My name is Feni Mirin. I am a citizen and native of West Papua, administrated and controlled by the Indonesian government.
I was born in Kono Village, in the West Papua on August 18th, 1994. I came to the US as a part of the student exchange program between the US and Indonesian government. I am so thankful of the opportunity not only to study for higher endeavors, but above all to experience first hand the uniqueness of living in a free-world. As a young adult, still in my growing years, I have nine (a) months of oppressions and inhumane treatment I had to endure in hands of the Indonesian military in 2013.
	The year 2013
I was in the 2nd year of my high school equivalent to (Grade II in US) when I had to live through the worst night-mare that life can offer especially to a young life in search of life. The experiences that have changed my being, and the way perceive live, that have also haunted my young life to this day. I always know that one day the same people that have tried to destroy my life and future will come back and take my life. Now, as a time is on my side, the opportunity avails I do believe that this is the time for me to find safety and refuge in the land where laws are followed and fully object.
I still can recall with a clear mind the day in the middle of our school classes and activities when unexpectedly the military stormed into our class with gun shots and war-shouts and forcing all of us to out in he open area, of our school. Everyone of us took to open areas, where we were pushed and kicked, pushed down on our knees or laid down on the ground. The whole school teachers and students were in panic mode. The military surrounded the school and there were no way anyone can escape.
Students kept in large groups, fearing our lives we were made to crawl while soldiers stepped on our backs with pressure. After hours of harassment the soldiers stated to forcefully dragged young students who they would suspect were part of the Papua Rebels of freedom fighters. I was caught in the rumble when 2 soldiers came to towards me with shouts and curse words ad stepped me on my face, dragged me with 17 other students taken into military custody. I was spit on my face, I was kicked on my groin, kicked on the legs, back and punched on my head. I saw blood all over the face of my colleagues. I was held by two military and made to run while they continued personnel to dish out punched and words.
	The month in prison
The nine months that I had spent in the hands the military were the most unbearable and inhumane treatment that I have ever experienced in my young life which had hunted me still to this very day.
It is the warmness of being free that I had felt living in this great country that had compelled me to seek help and assistance for my fears. Fear of living through that experiences again. The nine (a) month I had spent in the hands of the Indonesian military was aimed at running my young life. It was nine months of hunger and thirsts where we were denied food and water. It was nine months of being forced to go without sleep placed and dirty cold places. It was nine months of daily for have where hope to live for another day was lost. Nine months where I had to eat my tears ridiculed and filled with unbearable pains.